docker: initial copy from Buildroot
Do a RAW copy of the minimum number of files related to docker image
from Buildroot version 2023.02-rc3, and anddd a licence header to them.

NOTICE: this commit is not really usable to generate or run a docker
image, it is only meant for historical reasons, so the next commit(s) do
the adjusts needed to use these files in this project.

Signed-off-by: Ricardo Martincoski <ricardo.martincoski@gmail.com>