LICENSE: fix wrong GitLab license detection
Currently accessing the project license content at [1] wrongly shows:
 This project is licensed under the Affero General Public License v1.0.
even with the content starting with
                    GNU GENERAL PUBLIC LICENSE
                       Version 2, June 1991

By creating an empty project and adding the LICENSE file using the GUI
and using the GPLv2 template, the license detection works as expected.

Comparing the file in this repo with this one in the empty project we
can see that the section named "How to Apply These Terms to Your New
Programs" exists in the template.
So add it back to the LICENSE file.

It workarounds the wrong license detection and will show as expected:
 GNU General Public License v2.0 or later.
NOTICE the project is licensed under GPL-2.0-only but the LICENSE file
is the same for GPL-2.0-only and GPL-2.0-or-later projects.

At same time, try to clarify the license by:
- adding a COPYING file that references LICENSE and also display the
  SPDX license identifier;
- removing "(open source projects)" from the author name, since all
  projects that I publish are open source projects;
- adding electronic mail to the author name;
- using "version 2" instead of "version 2 only", following "Standard
  License Header" from [2];
- using the SPDX identifier at README;
- state at README that all files copied from toher projects can be
  redistributed under GPL-2.0;
- add license note to 'make help'.

[1] https://gitlab.com/RicardoMartincoski_opensource/sblkh/-/blob/main/LICENSE
[2] https://spdx.org/licenses/GPL-2.0-only.html

Signed-off-by: Ricardo Martincoski <ricardo.martincoski@gmail.com>